+++ tests/testthat/test-tmp_vars.R
@@ -1,6 +1,4 @@
-library(admiral.test)
-data(admiral_dm)
-dm <- select(admiral_dm, USUBJID)
+dm <- select(admiral.test::admiral_dm, USUBJID)
 
 # get_new_tmp_var ----
 ## Test 1: creating temporary variables works ----
@@ -54,4 +52,11 @@ test_that("remove_tmp_vars Test 6: removing temp variables works with the pipe o
       remove_tmp_vars()
   }
   expect_identical(colnames(dm), colnames(do_something_with_pipe(dm)))
+})
+
+
+## Test 7: running get_new_tmp_var on NULL dataset creates generic variable ----
+test_that("running get_new_tmp_var on NULL dataset creates generic variable", {
+  df <- NULL
+  expect_identical(get_new_tmp_var(df), sym("tmp_var_1"))
 })